feat: binarized the masks + lots of new metrics to fix
Former-commit-id: c840d14f722503d241f6bb6d899630ad6345aca0 [formerly e435a21234620add4f0e4e269a4141e5c1508cd9]
Former-commit-id: 8006af185fd68cc88b2305a02513106c16758d77

+++ src/utils/dataset.py
@@ -1,4 +1,3 @@
-import logging
 import os
 
 import numpy as np
@@ -24,6 +23,10 @@ class SphereDataset(Dataset):
         if self.transform is not None:
             augmentations = self.transform(image=image, mask=mask)
             image = augmentations["image"]
-            mask = augmentations["mask"].float()
+            mask = augmentations["mask"]
+
+        # make sure image and mask are floats
+        image = image.float()
+        mask = mask.float()
 
         return image, mask
